renamed TodayLine to TodayMarker
