+++ src/lib/markers/TimelineMarkersRenderer.js
@@ -2,7 +2,7 @@ import React from 'react'
 // import PropTypes from 'prop-types'
 import { TimelineMarkersConsumer } from './TimelineMarkersContext'
 import { TimelineMarkerType } from './markerType'
-import TodayLine from './implementations/TodayLine'
+import TodayMarker from './implementations/TodayMarker'
 
 /** Internal component used in timeline to render markers registered */
 const TimelineMarkersRenderer = () => {
@@ -12,7 +12,7 @@ const TimelineMarkersRenderer = () => {
         return markers.map(marker => {
           switch (marker.type) {
             case TimelineMarkerType.Today:
-              return <TodayLine key={marker.id} />
+              return <TodayMarker key={marker.id} />
             default:
               return null
           }
