+++ src/lib/markers/TodayMarker.js
