+++ src/lib/markers/implementations/TodayMarker.js
